fix: organise components on their own folders for consistency
