+++ packages/frontend/src/components/UserSettingsModal/OrganisationPanel/index.tsx
@@ -2,8 +2,8 @@ import { Button, FormGroup, InputGroup, Intent } from '@blueprintjs/core';
 import { ApiError } from 'common';
 import React, { FC, useEffect, useState } from 'react';
 import { useMutation, useQueryClient } from 'react-query';
-import { lightdashApi } from '../../api';
-import { useApp } from '../../providers/AppProvider';
+import { lightdashApi } from '../../../api';
+import { useApp } from '../../../providers/AppProvider';
 
 const updateOrgQuery = async (data: { organizationName: string }) =>
     lightdashApi<undefined>({
@@ -12,7 +12,7 @@ const updateOrgQuery = async (data: { organizationName: string }) =>
         body: JSON.stringify(data),
     });
 
-const OrganizationPanel: FC = () => {
+const OrganisationPanel: FC = () => {
     const queryClient = useQueryClient();
     const {
         errorLogs: { showError },
@@ -91,4 +91,4 @@ const OrganizationPanel: FC = () => {
     );
 };
 
-export default OrganizationPanel;
+export default OrganisationPanel;
