+++ packages/frontend/src/components/UserSettingsModal/ProfilePanel/index.tsx
@@ -2,8 +2,8 @@ import { Button, FormGroup, InputGroup, Intent } from '@blueprintjs/core';
 import { ApiError, LightdashUser, UpdateUserArgs, validateEmail } from 'common';
 import React, { FC, useEffect, useState } from 'react';
 import { useMutation, useQueryClient } from 'react-query';
-import { lightdashApi } from '../../api';
-import { useApp } from '../../providers/AppProvider';
+import { lightdashApi } from '../../../api';
+import { useApp } from '../../../providers/AppProvider';
 
 const updateUserQuery = async (data: UpdateUserArgs) =>
     lightdashApi<LightdashUser>({
