+++ packages/frontend/src/components/UserSettingsModal/ProjectManagementPanel/index.tsx
@@ -1,8 +1,8 @@
 import { ButtonGroup, Card, Classes } from '@blueprintjs/core';
 import { OrganizationProject } from 'common';
 import React, { FC } from 'react';
-import { useProjects } from '../../hooks/useProjects';
-import LinkButton from '../common/LinkButton';
+import { useProjects } from '../../../hooks/useProjects';
+import LinkButton from '../../common/LinkButton';
 
 const ProjectListItem: FC<{ project: OrganizationProject }> = ({
     project: { projectUuid, name },
